Split simulation module into submodules
